+++ src/util/terrain.c
@@ -675,7 +675,7 @@ main(int ac, char *av[])
     int arg_count;
     unsigned short *buf;
     int in_cookie, out_cookie;
-    int count;
+    size_t count;
     size_t ret;
 
     /* function to call to generate the terrain.  Default noise pattern is fbm */
@@ -722,7 +722,7 @@ main(int ac, char *av[])
     }
 
     ret = fwrite(buf, sizeof(*buf), count, stdout);
-    if (ret < (size_t)count)
+    if (ret < count)
 	perror("fwrite");
     bu_free(buf, "buf");
